cockpit.js: split location helpers to their own file
Move some location-related constants and helper functions to their own
file.  This is another clean move with no changes.
